Add building, buildingAd, buildingImage and related types to model. Added basic repositories and services implementation for the new types. Added mapper profiles Translate identity UI labels (cont.)

+++ HammadBroker.Data/Repositories/BuildingRepository.cs
@@ -7,7 +7,7 @@ using Microsoft.Extensions.Logging;
 
 namespace HammadBroker.Data.Repositories;
 
-public class BuildingRepository : Repository<DataContext, Building, long>, IBuildingRepository
+public class BuildingRepository : Repository<DataContext, Building, int>, IBuildingRepository
 {
 	/// <inheritdoc />
 	public BuildingRepository([NotNull] DataContext context, [NotNull] IConfiguration configuration, [NotNull] ILogger<BuildingRepository> logger)
